Logistic Regression of the Titanic Dataset
This is the process in building a Logistic Regression using the whole of the Titanic Dataset (the train and the test dataset).
Among other Classification models (K-Nearest Neighbor, Select Vector Machine, Random Forest, Decision Tree...) only Logistic Regression gave a better performance of model prediction when only the train dataset were used to predict, hence it was adopted in this model. 
This model used both train and test data to predict which passenger in the test data will survive or not. The model predicted 78% of the Passengers accurately and had 22% wrong prediction.

The data cleaning process is well explained in the Notebook and more preprocessing methods shall be adopted in earnest to increase the model prediction.

shall be explored to increase the model prediction as well.